+++ src/vdp_builtins.cpp
@@ -3,9 +3,9 @@
 namespace VDP {
 namespace Schema {
 Motor::Motor(std::string name, vex::motor &mot)
-    : Record(name), mot(mot), pos(new Double("Position(deg)")),
-      vel(new Double("velocity(dps)")), temp(new Uint8("Temperature(C)")),
-      voltage(new Double("Voltage(V)")), current(new Double("Current(%)")) {
+    : Record(name), mot(mot), pos(new Float("Position(deg)")),
+      vel(new Float("velocity(dps)")), temp(new Uint8("Temperature(C)")),
+      voltage(new Float("Voltage(V)")), current(new Float("Current(%)")) {
   Record::setFields({pos, vel, temp, voltage, current});
 }
